fix: use custom ctx object
